requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-CORS
pymysql
redis